store config xml to archive and load on server side
