+++ impl/src/main/java/ch/abertschi/sct/arquillian/client/DependencyResolver.java
@@ -16,6 +16,8 @@ public class DependencyResolver implements ProtocolArchiveProcessor
 
     private static final String SCT_IMPL = "ch.abertschi.sct:service-call-tracker-impl:" + "0.1.0-alpha7";
 
+    private static final String SCT_API = "ch.abertschi.sct:service-call-tracker-api:" + "0.1.0-alpha7";
+
     private static final String COMMONS_IO = "commons-io:commons-io:2.4";
 
     @Override
@@ -25,7 +27,7 @@ public class DependencyResolver implements ProtocolArchiveProcessor
         {
             LibraryContainer<?> container = (LibraryContainer<?>) archive;
             for (File f : ResolverUtil.get()
-                    .resolve(SCT_IMPL, COMMONS_IO).withTransitivity().asFile())
+                    .resolve(SCT_IMPL, SCT_API, COMMONS_IO).withTransitivity().asFile())
             {
                 LOG.info("arquillian service call tracker adding extension dependency to archive " + f.getAbsolutePath());
                 container.addAsLibraries(f);
